refactor: move all p2p mutations to mutations folder and refactor them to use the use-mutation hook
