+++ src/api/mutation/use-p2p-advertiser-payment-methods.tsx
@@ -1,0 +1,13 @@
+import { useMutation } from '../../base';
+import { AugmentedMutationOptions } from '../../base/use-mutation';
+
+export const useP2PAdvertiserPaymentMethods = ({
+    ...props
+}: Omit<AugmentedMutationOptions<'p2p_advertiser_payment_methods'>, 'name'> = {}) => {
+    const { data, ...rest } = useMutation({ name: 'p2p_advertiser_payment_methods', ...props });
+
+    return {
+        data: data?.p2p_advertiser_payment_methods,
+        ...rest,
+    };
+};
