+++ src/api/mutation/use-p2p-order-review.tsx
@@ -1,0 +1,11 @@
+import { useMutation } from '../../base';
+import { AugmentedMutationOptions } from '../../base/use-mutation';
+
+export const useP2POrderReview = ({ ...props }: Omit<AugmentedMutationOptions<'p2p_order_review'>, 'name'> = {}) => {
+    const { data, ...rest } = useMutation({ name: 'p2p_order_review', ...props });
+
+    return {
+        data: data?.p2p_order_review,
+        ...rest,
+    };
+};
